Add files via upload
This is a file of base64 patterns that represent strings that would be present if any of the functions in php7 were encoded to base64.  I'll probably add structure later by grouping them with their plain text translation.

This file is useful to swap out with patterns_raw.txt to gain additional insights into other strings to search for in base64.